refactor: move one shot into app.repository
